+++ .grenrc.js
@@ -8,7 +8,8 @@ module.exports = {
     "ignoreLabels": ["closed", "automation", "enhancement", "bug", "fix",
       "internal", "feature", "feat", "docs", "chore", "refactor", "ci",
       "perf", "test", "tests", "style", "groovy", "linux", "master", "mac", "windows",
-      "team:automation", "ready-to-merge", "backport-to-7.x"],
+      "team:automation", "ready-to-merge", "backport-to-7.x",
+      "backport-skip","dependencies", "java", "bump"],
     "groupBy": {
         "Enhancements": ["enhancement", "internal", "feature", "feat"],
         "Bug Fixes": ["bug", "fix"],