+++ src/test/groovy/ApmBasePipelineTest.groovy
@@ -53,7 +53,7 @@ class ApmBasePipelineTest extends DeclarativePipelineTest {
     SECRET_CODECOV('secret-codecov'), SECRET_ERROR('secretError'),
     SECRET_NAME('secret/team/ci/secret-name'), SECRET_NOT_VALID('secretNotValid'), SECRET_GITHUB_APP('secret/observability-team/ci/github-app'),
     SECRET_NPMJS('secret/apm-team/ci/elastic-observability-npmjs'), SECRET_NPMRC('secret-npmrc'),
-    SECRET_TOTP('secret-totp'), SECRET_GCP('service-account/apm-rum-admin')
+    SECRET_TOTP('secret-totp'), SECRET_GCP('service-account/apm-rum-admin'), SECRET_GCP_PROVISIONER('service-account/provisioner')
 
     VaultSecret(String value) {
       this.value = value
@@ -323,6 +323,7 @@ class ApmBasePipelineTest extends DeclarativePipelineTest {
     helper.registerAllowedMethod('readJSON', [Map.class], { m ->
       return readJSON(m)
     })
+    helper.registerAllowedMethod('readProperties', [Map.class], { [:] })
     helper.registerAllowedMethod('retry', [Integer.class, Closure.class], { count, c ->
       Exception lastError = null
       while (count-- > 0) {
@@ -357,7 +358,9 @@ class ApmBasePipelineTest extends DeclarativePipelineTest {
     helper.registerAllowedMethod('withEnv', [List.class, Closure.class], TestUtils.withEnvInterceptor)
     helper.registerAllowedMethod('wrap', [Map.class, Closure.class], TestUtils.wrapInterceptor)
     helper.registerAllowedMethod('writeFile', [Map.class], { m ->
-      (new File("target/${m.file}")).withWriter('UTF-8') { writer ->
+      File f = new File("target/${m.file}")
+      f.getParentFile().mkdirs()
+      f.withWriter('UTF-8') { writer ->
         writer.write(m.text)
       }
     })
@@ -391,6 +394,14 @@ class ApmBasePipelineTest extends DeclarativePipelineTest {
     })
     helper.registerAllowedMethod('dockerImageExists', [Map.class], { true })
     helper.registerAllowedMethod('dockerLogin', [Map.class], { true })
+    helper.registerAllowedMethod('downloadWithCurl', [Map.class], { m ->
+      def script = loadScript('vars/downloadWithCurl.groovy')
+      return script.call(m)
+    })
+    helper.registerAllowedMethod('downloadWithWget', [Map.class], { m ->
+      def script = loadScript('vars/downloadWithWget.groovy')
+      return script.call(m)
+    })
     helper.registerAllowedMethod('echoColor', [Map.class], { m ->
       def echoColor = loadScript('vars/echoColor.groovy')
       echoColor.call(m)
@@ -583,6 +594,10 @@ class ApmBasePipelineTest extends DeclarativePipelineTest {
       def script = loadScript('vars/withGCPEnv.groovy')
       return script.call(m, c)
     })
+    helper.registerAllowedMethod('withGhEnv', [Map.class, Closure.class], { m, c ->
+      def script = loadScript('vars/withGhEnv.groovy')
+      return script.call(m, c)
+    })
     helper.registerAllowedMethod('withGithubNotify', [Map.class, Closure.class], null)
     helper.registerAllowedMethod('withGithubCheck', [Map.class, Closure.class], { m, body -> body() })
     helper.registerAllowedMethod('withGithubStatus', [Map.class, Closure.class], { m, body -> body() })
@@ -599,6 +614,13 @@ class ApmBasePipelineTest extends DeclarativePipelineTest {
     helper.registerAllowedMethod('withMageEnv', [Closure.class], { c ->
       def script = loadScript('vars/withMageEnv.groovy')
       return script.call(c)
+    })
+    helper.registerAllowedMethod('withNodeJSEnv', [Map.class, Closure.class], { m, c ->
+      def script = loadScript('vars/withNodeJSEnv.groovy')
+      return script.call(m, c)
+    })
+    helper.registerAllowedMethod('withNodeJSEnvUnix', [Map.class, Closure.class], { m, c ->
+      return true
     })
   }
 
@@ -650,6 +672,9 @@ class ApmBasePipelineTest extends DeclarativePipelineTest {
     }
     if(VaultSecret.SECRET_GCP.equals(s)){
       return [data: [ value: 'mytoken' ]]
+    }
+    if(VaultSecret.SECRET_GCP_PROVISIONER.equals(s)){
+      return [data: [ credentials: 'my_json_credentials' ]]
     }
     if(VaultSecret.SECRET_GITHUB_APP.equals(s)){
       return [data: [ key: new File('src/test/resources/github-app-private-key-tests.pem').text, installation_id: '123', app_id: '42' ]]