+++ src/test/groovy/GithubPullRequestsStepTests.groovy
@@ -30,13 +30,6 @@ class GithubPullRequestsStepTests extends ApmBasePipelineTest {
   }
 
   @Test
-  void test_windows() throws Exception {
-    testWindows() {
-      script.call()
-    }
-  }
-
-  @Test
   void test_with_no_data() throws Exception {
     helper.registerAllowedMethod('gh', [Map.class], { return '' })
     def ret = script.call()