+++ vars/bumpUtils.groovy
@@ -40,3 +40,90 @@ def prepareContext(Map args = [:]) {
       branch: args.branchName,
       credentialsId: args.get('credentialsId', '2a9602aa-ab9f-4e52-baf3-b71ca88469c7-UserAndToken'))
 }
+
+def parseArguments(Map args = [:]) {
+  def arguments = [
+      title: "${args.title}", labels: "${args.labels}", description: "${args.message}", base: "${args.branchName}"
+  ]
+  if (args.assign?.trim()) {
+    arguments['assign'] = args.assign
+  }
+  if (args.reviewer?.trim()) {
+    arguments['reviewer'] = args.reviewer
+  }
+  return arguments
+}
+
+def getCurrentMinorReleaseFor8() {
+  return getValueForPropertyKey(current8Key())
+}
+
+def getNextMinorReleaseFor8() {
+  return getValueForPropertyKey(nextMinor8Key())
+}
+
+def getNextPatchReleaseFor8() {
+  return getValueForPropertyKey(nextPatch8Key())
+}
+
+def getCurrentMinorReleaseFor7() {
+  return getValueForPropertyKey(current7Key())
+}
+
+def getNextMinorReleaseFor7() {
+  return getValueForPropertyKey(nextMinor7Key())
+}
+
+def getNextPatchReleaseFor7() {
+  return getValueForPropertyKey(nextPatch7Key())
+}
+
+def getCurrentMinorReleaseFor6() {
+  return getValueForPropertyKey(current6Key())
+}
+
+// for internal purposes
+def getValueForPropertyKey(String key) {
+  def stackVersions = readProperties(text: libraryResource('versions/releases.properties'))
+  def version = stackVersions[key]
+  if (!version?.trim()) {
+    error("getValueForPropertyKey: '${key}' has an empty value")
+  }
+  return version
+}
+
+def current8Key() {
+  return 'current_8'
+}
+
+def nextMinor8Key() {
+  return 'next_minor_8'
+}
+
+def nextPatch8Key() {
+  return 'next_patch_8'
+}
+
+def current7Key() {
+  return 'current_7'
+}
+
+def nextMinor7Key() {
+  return 'next_minor_7'
+}
+
+def nextPatch7Key() {
+  return 'next_patch_7'
+}
+
+def current6Key() {
+  return 'current_6'
+}
+
+def areStackVersionsAvailable(stackVersions) {
+  // TODO: to support the 8.x branches
+  return isVersionAvailable(stackVersions.get(current6Key())) &&
+    isVersionAvailable(stackVersions.get(current7Key())) &&
+    isVersionAvailable(stackVersions.get(nextMinor7Key())) &&
+    isVersionAvailable(stackVersions.get(nextPatch7Key()))
+}